% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf.R
\name{calc_npv}
\alias{calc_npv}
\title{Net Present Value}
\usage{
calc_npv(cf, r, c0 = 0)
}
\arguments{
\item{cf}{Numeric. Cash flow for each period.}

\item{r}{Numeric. Cost of capital.}

\item{c0}{Numeric. Initial investment. Default 0.}
}
\value{
Rate of return.
}
\description{
Net Present Value
}
\examples{
cf <- rep(25000, 60)
npv <- calc_npv(cf, 0.0064, 1000000)
}
