% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf.R
\name{calc_pv}
\alias{calc_pv}
\title{Present Value}
\usage{
calc_pv(fv, r, n)
}
\arguments{
\item{fv}{Numeric. Future value.}

\item{r}{Numeric. Rate of return.}

\item{n}{Numeric. Number of periods.}
}
\value{
The calculated present value.
}
\description{
Present Value
}
\examples{
pv <- calc_pv(2000, 0.03, 3)
}
