% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf.R
\name{dcf_sim_norm}
\alias{dcf_sim_norm}
\title{Monte Carlo simulation for unlevered free cash flow model with perpetuity growth method}
\usage{
dcf_sim_norm(
  n_sim,
  CF,
  future_value,
  net_debt,
  shares,
  wacc_mu,
  wacc_sigma,
  g_mu,
  g_sigma,
  years,
  seed = NULL
)
}
\arguments{
\item{n_sim}{}

\item{CF}{}

\item{net_debt}{}

\item{shares}{}

\item{wacc_mu}{}

\item{wacc_sigma}{}

\item{g_mu}{}

\item{g_sigma}{}

\item{years}{}

\item{seed}{}

\item{FV}{}
}
\value{
Simulated valuations for company.
}
\description{
Monte Carlo simulation for unlevered free cash flow model with perpetuity growth method
}
