% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm_sim.R
\name{ddm_sim}
\alias{ddm_sim}
\title{Monte Carlo Simulation for the dividend discount model (Gordon Growth model)}
\usage{
ddm_sim(
  d,
  r,
  g_mu = NULL,
  g_sigma = NULL,
  g_min = NULL,
  g_max = NULL,
  n_sim = 1000,
  seed = NULL,
  distribution = c("normal", "triangle", "uniform")
)
}
\arguments{
\item{d}{Numeric. Estimated value of dividend in the next year.}

\item{r}{Numeric. Constant cost of equity.}

\item{g_mu}{Numeric. The mean of the normal distribution.}

\item{g_sigma}{Numeric.The standard deviation of the normal distribution.}

\item{g_min}{Numeric. Lower limit of the distribution.}

\item{g_max}{Numeric. Upper limit of the distribution.}

\item{n_sim}{Numeric. The number of simulations.}

\item{seed}{Numeric. State for random number generation.}

\item{distribution}{Character. The type of distribution.}
}
\value{
A numeric vector with the current simulated stock prices.
}
\description{
Monte Carlo Simulation for the dividend discount model (Gordon Growth model)
}
\examples{
stock_prices <- ddm_sim(1.5, 0.05, g_mu = 0.03, g_sigma = 0.01)
}
