% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{growth_rate}
\alias{growth_rate}
\title{Dividend growth rate}
\usage{
growth_rate(roe, b)
}
\arguments{
\item{roe}{Numeric. Return on equity.}

\item{b}{Numeric. Plowback ratio.}
}
\value{
The company's growth rate.
}
\description{
Growth rate equals the return on equity times the plowback ratio, or growth
is determined by how much of earnings is put back into the firm, and how
profitable those earnings are.
}
\examples{
g <- growth_rate(0.25, 0.4)
}
