% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wacc.R
\name{load_beta}
\alias{load_beta}
\title{Load total beta data}
\usage{
load_beta(
  region = c("US", "Europe", "Japan", "China", "India", "Emerging Markets", "Global")
)
}
\arguments{
\item{region}{}
}
\description{
\code{load_beta} loads Aswath Damodaran's latest total beta by industry
data.
}
