% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wacc.R
\name{load_beta_data}
\alias{load_beta_data}
\title{Load total beta data}
\usage{
load_beta_data(region = c("US", "Europe", "Japan", "China", "India", "Global"))
}
\arguments{
\item{region}{}
}
\value{

}
\description{
\code{load_beta_data} loads Aswath Damodaran's latest total beta by industry
data.
}
