% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wacc.R
\name{load_eps_growth_rate}
\alias{load_eps_growth_rate}
\title{Load fundamental growth rate in EPS data}
\usage{
load_eps_growth_rate(
  region = c("US", "Europe", "Japan", "Emerging Markets", "Global")
)
}
\arguments{
\item{region}{}
}
\description{
\code{load_eps_growth_rate} loads Aswath Damodaran's latest fundamental growth
rate in EPS by industry data.
}
