% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wacc.R
\name{load_wacc}
\alias{load_wacc}
\title{Load cost of capital data}
\usage{
load_wacc(
  region = c("US", "Europe", "Japan", "China", "India", "Emerging Markets", "Global")
)
}
\arguments{
\item{region}{}
}
\description{
\code{load_wacc} loads Aswath Damodaran's latest cost of capital by industry
data.
}
