% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wacc.R
\name{load_wacc_data}
\alias{load_wacc_data}
\title{Load cost of capital data}
\usage{
load_wacc_data(region = c("US", "Europe", "Japan", "China", "India", "Global"))
}
\arguments{
\item{region}{}
}
\value{

}
\description{
\code{load_wacc_data} loads Aswath Damodaran's latest cost of capital by industry
data.
}
