% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{plowback_ratio}
\alias{plowback_ratio}
\title{Dividend plowback ratio.}
\usage{
plowback_ratio(d, eps)
}
\arguments{
\item{d}{Numeric. Dividend per share.}

\item{eps}{Numeric. Earnings per share.}
}
\value{

}
\description{
Dividend plowback ratio.
}
\examples{
b <- plowback_ratio(0.84, 5.73)
}
