% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm_sim.R
\name{ddm}
\alias{ddm}
\title{Gordon Growth Model}
\usage{
ddm(d, r, g)
}
\arguments{
\item{d}{Numeric. Estimated value of dividend in the next year.}

\item{r}{Numeric. Constant cost of equity.}

\item{g}{Numeric. Constant growth rate for dividends in perpetuity.}
}
\value{
The estimated price per share.
}
\description{
Gordon Growth Model
}
\examples{
stock_price <- ddm(1.5, 0.05, 0.03)
}
