% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm_sim.R
\name{ddm_sim}
\alias{ddm_sim}
\title{Monte Carlo Simulation for the dividend discount model (Gordon Growth model)}
\usage{
ddm_sim(
  price,
  dividend,
  g_mu = NULL,
  g_sigma = NULL,
  g_min = NULL,
  g_max = NULL,
  n_sim = 1000,
  seed = NULL,
  distribution = c("normal", "triangle", "uniform")
)
}
\arguments{
\item{price}{The stock price of the company.}

\item{dividend}{The distributed dividend for the stock.}

\item{g_mu}{The mean of the normal distribution.}

\item{g_sigma}{The standard deviation of the normal distribution.}

\item{g_min}{Lower limit of the distribution.}

\item{g_max}{Upper limit of the distribution.}

\item{n_sim}{The number of simulations.}

\item{seed}{State for random number generation.}

\item{distribution}{The type of distribution.}
}
\value{
A numeric vector with the current simulated stock prices.
}
\description{
Monte Carlo Simulation for the dividend discount model (Gordon Growth model)
}
\examples{
dividends <- ddm_sim(25.0, 1.5, g_mu = 0.03, g_sigma = 0.01)
}
